Fixed another critical server error
