+++ src/main/java/dev/lazurite/rayon/api/packet/RayonSpawnS2CPacket.java
@@ -5,14 +5,9 @@ import dev.lazurite.rayon.physics.body.EntityRigidBody;
 import dev.lazurite.rayon.physics.helper.math.QuaternionHelper;
 import dev.lazurite.rayon.physics.helper.math.VectorHelper;
 import dev.lazurite.rayon.util.exception.RayonSpawnException;
-import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.fabricmc.fabric.api.networking.v1.PacketByteBufs;
-import net.fabricmc.fabric.api.networking.v1.PacketSender;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.network.ClientPlayNetworkHandler;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.EntityType;
 import net.minecraft.network.Packet;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.util.Identifier;
@@ -20,7 +15,6 @@ import net.minecraft.util.registry.Registry;
 
 import javax.vecmath.Quat4f;
 import javax.vecmath.Vector3f;
-import java.util.UUID;
 
 /**
  * This custom spawn packet can only be used with physics entities. It isn't required in order to
@@ -33,39 +27,7 @@ import java.util.UUID;
 public class RayonSpawnS2CPacket {
     public static final Identifier PACKET_ID = new Identifier(Rayon.MODID, "rayon_spawn_s2c_packet");
 
-    public static void accept(MinecraftClient client, ClientPlayNetworkHandler handler, PacketByteBuf buf, PacketSender sender) {
-        EntityType<?> entityType = Registry.ENTITY_TYPE.get(buf.readVarInt());
-        int entityId = buf.readInt();
-        UUID uuid = buf.readUuid();
-
-        Quat4f orientation = QuaternionHelper.fromBuffer(buf);
-        Vector3f position = VectorHelper.fromBuffer(buf);
-        Vector3f linearVelocity = VectorHelper.fromBuffer(buf);
-        Vector3f angularVelocity = VectorHelper.fromBuffer(buf);
-
-        client.execute(() -> {
-            Entity entity = entityType.create(client.world);
-            entity.setEntityId(entityId);
-            entity.setUuid(uuid);
-
-            float x = position.x;
-            float y = position.y - (float) entity.getBoundingBox().getYLength() / 2.0f;
-            float z = position.z;
-            entity.updatePosition(x, y, z);
-            entity.pitch = QuaternionHelper.getPitch(orientation);
-            entity.yaw = QuaternionHelper.getYaw(orientation);
-
-            EntityRigidBody body = EntityRigidBody.get(entity);
-            body.setOrientation(orientation);
-            body.setPosition(position);
-            body.setLinearVelocity(linearVelocity);
-            body.setAngularVelocity(angularVelocity);
-
-            client.world.addEntity(entityId, entity);
-        });
-    }
-
-    public static Packet<?> get(Entity entity) {
+    public static Packet<?> get (Entity entity){
         if (!EntityRigidBody.is(entity)) {
             throw new RayonSpawnException("The given entity is not registered.");
         }
@@ -88,9 +50,5 @@ public class RayonSpawnS2CPacket {
         VectorHelper.toBuffer(buf, body.getAngularVelocity(new Vector3f()));
 
         return ServerPlayNetworking.createS2CPacket(PACKET_ID, buf);
-    }
-
-    public static void register() {
-        ClientPlayNetworking.registerGlobalReceiver(PACKET_ID, RayonSpawnS2CPacket::accept);
     }
 }
